# Remove corech options, redundant tests