--- a/.gitignore
+++ b/.gitignore
@@ -46,7 +46,6 @@ Thumbs.db
 /example/build/
 /test/data/monaco.osrm*
 /test/data/ch
-/test/data/corech
 /test/data/mld
 /cmake/postinst