--- a/features/options/datastore/datastore.feature
+++ b/features/options/datastore/datastore.feature
@@ -33,10 +33,6 @@ Feature: osrm-datastore command line options
         When I try to run "osrm-datastore {processed_file} --dataset-name cucumber/only_metric_test --only-metric"
         Then it should exit successfully
 
-    Scenario: osrm-datastore - Displaying help should work
-        When I try to run "osrm-datastore {processed_file} --help"
-        Then it should exit successfully
-
     Scenario: osrm-datastore - Errors on invalid path
         When I try to run "osrm-datastore invalid_path.osrm"
         Then stderr should contain "[error] Config contains invalid file paths."