- breaking change: supplying `--corech` parameter to `osrm-extract` now leads to an error
- associated tests removed
- irrelevant tests on help output removed


<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1156.63<br/>plain u32: 1155.77<br/>aliased double: 1224.5<br/>plain double: 1145.03 | aliased u32: 1115.3<br/>plain u32: 1107.27<br/>aliased double: 975.774<br/>plain double: 970.021 |
| e2e_match_ch | Ops: 40.26 ± 0.12 ops/s. Best: 40.40 ops/s<br/>Total: 3254.03ms ± 9.96ms. Best: 3242.83ms<br/>Min time: 2.15ms ± 0.05ms<br/>Mean time: 24.84ms ± 0.08ms<br/>Median time: 18.63ms ± 0.12ms<br/>95th percentile: 86.32ms ± 0.42ms<br/>99th percentile: 104.54ms ± 0.44ms<br/>Max time: 112.90ms ± 0.83ms | Ops: 39.47 ± 0.11 ops/s. Best: 39.57 ops/s<br/>Total: 3319.18ms ± 8.57ms. Best: 3310.18ms<br/>Min time: 2.24ms ± 0.04ms<br/>Mean time: 25.34ms ± 0.07ms<br/>Median time: 18.87ms ± 0.16ms<br/>95th percentile: 88.96ms ± 0.41ms<br/>99th percentile: 107.63ms ± 0.24ms<br/>Max time: 116.37ms ± 0.39ms |
| e2e_match_mld | Ops: 63.60 ± 0.19 ops/s. Best: 63.79 ops/s<br/>Total: 2059.86ms ± 6.40ms. Best: 2053.49ms<br/>Min time: 1.80ms ± 0.02ms<br/>Mean time: 15.72ms ± 0.05ms<br/>Median time: 8.40ms ± 0.17ms<br/>95th percentile: 51.92ms ± 0.28ms<br/>99th percentile: 60.51ms ± 0.27ms<br/>Max time: 69.90ms ± 0.40ms | Ops: 62.90 ± 0.16 ops/s. Best: 63.22 ops/s<br/>Total: 2082.99ms ± 5.19ms. Best: 2072.22ms<br/>Min time: 1.81ms ± 0.02ms<br/>Mean time: 15.90ms ± 0.04ms<br/>Median time: 8.57ms ± 0.15ms<br/>95th percentile: 52.34ms ± 0.10ms<br/>99th percentile: 62.04ms ± 0.56ms<br/>Max time: 71.33ms ± 0.15ms |
| e2e_nearest_ch | Ops: 852.76 ± 4.71 ops/s. Best: 862.44 ops/s<br/>Total: 1172.70ms ± 7.14ms. Best: 1159.50ms<br/>Min time: 0.99ms ± 0.01ms<br/>Mean time: 1.17ms ± 0.01ms<br/>Median time: 1.08ms ± 0.00ms<br/>95th percentile: 1.64ms ± 0.01ms<br/>99th percentile: 1.70ms ± 0.01ms<br/>Max time: 7.16ms ± 3.52ms | Ops: 853.24 ± 8.76 ops/s. Best: 863.70 ops/s<br/>Total: 1171.70ms ± 12.21ms. Best: 1157.81ms<br/>Min time: 0.98ms ± 0.01ms<br/>Mean time: 1.17ms ± 0.01ms<br/>Median time: 1.08ms ± 0.00ms<br/>95th percentile: 1.62ms ± 0.01ms<br/>99th percentile: 1.71ms ± 0.04ms<br/>Max time: 6.40ms ± 3.10ms |
| e2e_nearest_mld | Ops: 857.99 ± 3.24 ops/s. Best: 863.54 ops/s<br/>Total: 1165.60ms ± 4.74ms. Best: 1158.03ms<br/>Min time: 0.99ms ± 0.00ms<br/>Mean time: 1.17ms ± 0.00ms<br/>Median time: 1.08ms ± 0.00ms<br/>95th percentile: 1.62ms ± 0.01ms<br/>99th percentile: 1.67ms ± 0.01ms<br/>Max time: 7.05ms ± 3.46ms | Ops: 856.52 ± 2.73 ops/s. Best: 861.47 ops/s<br/>Total: 1167.45ms ± 3.78ms. Best: 1160.80ms<br/>Min time: 0.99ms ± 0.00ms<br/>Mean time: 1.17ms ± 0.00ms<br/>Median time: 1.08ms ± 0.00ms<br/>95th percentile: 1.62ms ± 0.01ms<br/>99th percentile: 1.68ms ± 0.01ms<br/>Max time: 6.77ms ± 3.35ms |
| e2e_route_ch | Ops: 335.30 ± 1.55 ops/s. Best: 338.03 ops/s<br/>Total: 2982.09ms ± 13.77ms. Best: 2958.32ms<br/>Min time: 1.26ms ± 0.02ms<br/>Mean time: 2.98ms ± 0.01ms<br/>Median time: 3.01ms ± 0.01ms<br/>95th percentile: 3.99ms ± 0.03ms<br/>99th percentile: 4.46ms ± 0.05ms<br/>Max time: 7.70ms ± 2.87ms | Ops: 331.44 ± 5.63 ops/s. Best: 343.15 ops/s<br/>Total: 3016.93ms ± 52.89ms. Best: 2914.20ms<br/>Min time: 1.26ms ± 0.03ms<br/>Mean time: 3.02ms ± 0.06ms<br/>Median time: 3.04ms ± 0.05ms<br/>95th percentile: 4.04ms ± 0.09ms<br/>99th percentile: 4.52ms ± 0.07ms<br/>Max time: 8.81ms ± 4.17ms |
| e2e_route_mld | Ops: 275.74 ± 4.32 ops/s. Best: 281.35 ops/s<br/>Total: 3626.94ms ± 58.65ms. Best: 3554.35ms<br/>Min time: 1.26ms ± 0.02ms<br/>Mean time: 3.63ms ± 0.06ms<br/>Median time: 3.68ms ± 0.06ms<br/>95th percentile: 5.00ms ± 0.09ms<br/>99th percentile: 5.59ms ± 0.13ms<br/>Max time: 8.49ms ± 2.13ms | Ops: 271.01 ± 4.76 ops/s. Best: 276.97 ops/s<br/>Total: 3688.02ms ± 62.59ms. Best: 3610.51ms<br/>Min time: 1.27ms ± 0.02ms<br/>Mean time: 3.69ms ± 0.06ms<br/>Median time: 3.73ms ± 0.06ms<br/>95th percentile: 5.10ms ± 0.07ms<br/>99th percentile: 5.69ms ± 0.12ms<br/>Max time: 8.69ms ± 2.44ms |
| e2e_table_ch | Ops: 312.02 ± 2.50 ops/s. Best: 318.11 ops/s<br/>Total: 3204.64ms ± 25.58ms. Best: 3143.57ms<br/>Min time: 1.66ms ± 0.02ms<br/>Mean time: 3.21ms ± 0.03ms<br/>Median time: 3.21ms ± 0.03ms<br/>95th percentile: 4.46ms ± 0.04ms<br/>99th percentile: 4.81ms ± 0.04ms<br/>Max time: 9.34ms ± 3.05ms | Ops: 308.87 ± 1.65 ops/s. Best: 311.76 ops/s<br/>Total: 3237.76ms ± 18.09ms. Best: 3207.60ms<br/>Min time: 1.73ms ± 0.02ms<br/>Mean time: 3.24ms ± 0.02ms<br/>Median time: 3.24ms ± 0.02ms<br/>95th percentile: 4.51ms ± 0.03ms<br/>99th percentile: 4.86ms ± 0.05ms<br/>Max time: 8.97ms ± 2.64ms |
| e2e_table_mld | Ops: 100.32 ± 0.19 ops/s. Best: 100.64 ops/s<br/>Total: 9967.85ms ± 19.43ms. Best: 9936.00ms<br/>Min time: 4.01ms ± 0.11ms<br/>Mean time: 9.97ms ± 0.02ms<br/>Median time: 9.93ms ± 0.07ms<br/>95th percentile: 15.16ms ± 0.04ms<br/>99th percentile: 16.00ms ± 0.07ms<br/>Max time: 19.08ms ± 2.92ms | Ops: 102.05 ± 0.72 ops/s. Best: 103.27 ops/s<br/>Total: 9799.91ms ± 71.15ms. Best: 9683.63ms<br/>Min time: 3.99ms ± 0.02ms<br/>Mean time: 9.80ms ± 0.07ms<br/>Median time: 9.76ms ± 0.10ms<br/>95th percentile: 14.97ms ± 0.09ms<br/>99th percentile: 15.79ms ± 0.14ms<br/>Max time: 19.37ms ± 2.90ms |
| e2e_trip_ch | Ops: 92.19 ± 0.97 ops/s. Best: 93.54 ops/s<br/>Total: 10850.07ms ± 113.09ms. Best: 10690.59ms<br/>Min time: 1.69ms ± 0.15ms<br/>Mean time: 10.85ms ± 0.12ms<br/>Median time: 10.28ms ± 0.12ms<br/>95th percentile: 19.18ms ± 0.14ms<br/>99th percentile: 21.35ms ± 0.28ms<br/>Max time: 24.70ms ± 2.46ms | Ops: 89.73 ± 0.36 ops/s. Best: 90.05 ops/s<br/>Total: 11145.05ms ± 46.61ms. Best: 11105.46ms<br/>Min time: 1.80ms ± 0.14ms<br/>Mean time: 11.15ms ± 0.05ms<br/>Median time: 10.62ms ± 0.05ms<br/>95th percentile: 19.59ms ± 0.09ms<br/>99th percentile: 21.45ms ± 0.08ms<br/>Max time: 23.48ms ± 0.23ms |
| e2e_trip_mld | Ops: 54.35 ± 0.49 ops/s. Best: 54.96 ops/s<br/>Total: 18407.21ms ± 170.13ms. Best: 18194.52ms<br/>Min time: 1.95ms ± 0.28ms<br/>Mean time: 18.40ms ± 0.18ms<br/>Median time: 17.96ms ± 0.17ms<br/>95th percentile: 29.78ms ± 0.24ms<br/>99th percentile: 32.23ms ± 0.42ms<br/>Max time: 34.33ms ± 0.76ms | Ops: 54.59 ± 0.43 ops/s. Best: 55.30 ops/s<br/>Total: 18327.56ms ± 140.51ms. Best: 18083.69ms<br/>Min time: 1.91ms ± 0.27ms<br/>Mean time: 18.32ms ± 0.15ms<br/>Median time: 17.82ms ± 0.15ms<br/>95th percentile: 29.69ms ± 0.18ms<br/>99th percentile: 32.15ms ± 0.32ms<br/>Max time: 35.26ms ± 0.92ms |
| json-render | String: 5.95386ms<br/>Stringstream: 9.48053ms<br/>Vector: 6.97761ms | String: 5.90178ms<br/>Stringstream: 9.60861ms<br/>Vector: 7.11491ms |
| match_ch | Default radius: <br/>4.65888ms/req at 82 coordinate<br/>0.0568156ms/coordinate<br/>Radius 10m: <br/>16.5364ms/req at 82 coordinate<br/>0.201664ms/coordinate | Default radius: <br/>4.65008ms/req at 82 coordinate<br/>0.0567083ms/coordinate<br/>Radius 10m: <br/>16.2529ms/req at 82 coordinate<br/>0.198206ms/coordinate |
| match_mld | Default radius: <br/>3.00522ms/req at 82 coordinate<br/>0.0366491ms/coordinate<br/>Radius 10m: <br/>11.4487ms/req at 82 coordinate<br/>0.139618ms/coordinate | Default radius: <br/>3.09269ms/req at 82 coordinate<br/>0.0377157ms/coordinate<br/>Radius 10m: <br/>11.2672ms/req at 82 coordinate<br/>0.137405ms/coordinate |
| osrm_contract | Time: 103.73s Peak RAM: 201.96MB | Time: 103.45s Peak RAM: 202.78MB |
| osrm_customize | Time: 1.30s Peak RAM: 117.78MB | Time: 1.30s Peak RAM: 117.53MB |
| osrm_extract | Time: 11.96s Peak RAM: 429.17MB | Time: 12.01s Peak RAM: 430.17MB |
| osrm_partition | Time: 2.22s Peak RAM: 140.71MB | Time: 2.14s Peak RAM: 152.02MB |
| packedvector | random write:<br/>std::vector 10715.1 ms<br/>util::packed_vector 74231.8 ms<br/>slowdown: 6.9278<br/>random read:<br/>std::vector 8937.86 ms<br/>util::packed_vector 30601.1 ms<br/>slowdown: 3.42376 | random write:<br/>std::vector 10202.2 ms<br/>util::packed_vector 74704.6 ms<br/>slowdown: 7.32241<br/>random read:<br/>std::vector 8965.1 ms<br/>util::packed_vector 30873.7 ms<br/>slowdown: 3.44377 |
| random_match_ch | 500 matches, default radius<br/>ops: 194.50 ± 3.93 ops/s. best: 199.08ops/s.<br/>total: 293.21 ± 5.99ms. best: 286.32ms.<br/>avg: 5.14 ± 0.11ms<br/>min: 0.14 ± 0.01ms<br/>max: 29.94 ± 3.07ms<br/>p99: 29.94 ± 3.07ms<br/><br/>500 matches, radius=10<br/>ops: 58.74 ± 0.07 ops/s. best: 58.81ops/s.<br/>total: 1089.57 ± 1.22ms. best: 1088.20ms.<br/>avg: 17.02 ± 0.02ms<br/>min: 0.15 ± 0.01ms<br/>max: 248.86 ± 0.46ms<br/>p99: 248.86 ± 0.46ms<br/><br/>500 matches, radius=20<br/>ops: 14.23 ± 0.03 ops/s. best: 14.27ops/s.<br/>total: 4568.29 ± 8.20ms. best: 4554.64ms.<br/>avg: 70.28 ± 0.13ms<br/>min: 0.30 ± 0.00ms<br/>max: 1249.77 ± 3.51ms<br/>p99: 1249.77 ± 3.51ms | 500 matches, default radius<br/>ops: 200.11 ± 0.72 ops/s. best: 200.86ops/s.<br/>total: 284.84 ± 1.03ms. best: 283.79ms.<br/>avg: 5.00 ± 0.02ms<br/>min: 0.14 ± 0.01ms<br/>max: 28.01 ± 0.06ms<br/>p99: 28.01 ± 0.06ms<br/><br/>500 matches, radius=10<br/>ops: 59.63 ± 0.14 ops/s. best: 59.82ops/s.<br/>total: 1073.27 ± 2.44ms. best: 1069.87ms.<br/>avg: 16.77 ± 0.04ms<br/>min: 0.15 ± 0.00ms<br/>max: 234.29 ± 0.86ms<br/>p99: 234.29 ± 0.86ms<br/><br/>500 matches, radius=20<br/>ops: 14.48 ± 0.05 ops/s. best: 14.54ops/s.<br/>total: 4489.78 ± 15.83ms. best: 4469.51ms.<br/>avg: 69.07 ± 0.24ms<br/>min: 0.30 ± 0.01ms<br/>max: 1159.11 ± 4.88ms<br/>p99: 1159.11 ± 4.88ms |
| random_match_mld | 500 matches, default radius<br/>ops: 305.89 ± 1.70 ops/s. best: 307.38ops/s.<br/>total: 186.35 ± 1.04ms. best: 185.44ms.<br/>avg: 3.27 ± 0.02ms<br/>min: 0.13 ± 0.00ms<br/>max: 18.91 ± 0.06ms<br/>p99: 18.91 ± 0.06ms<br/><br/>500 matches, radius=10<br/>ops: 109.11 ± 0.21 ops/s. best: 109.39ops/s.<br/>total: 586.55 ± 1.12ms. best: 585.08ms.<br/>avg: 9.16 ± 0.02ms<br/>min: 0.14 ± 0.00ms<br/>max: 109.76 ± 0.35ms<br/>p99: 109.76 ± 0.35ms<br/><br/>500 matches, radius=20<br/>ops: 22.26 ± 0.10 ops/s. best: 22.44ops/s.<br/>total: 2919.79 ± 13.68ms. best: 2896.52ms.<br/>avg: 44.92 ± 0.21ms<br/>min: 0.19 ± 0.00ms<br/>max: 576.44 ± 9.19ms<br/>p99: 576.44 ± 9.19ms | 500 matches, default radius<br/>ops: 306.22 ± 1.79 ops/s. best: 307.93ops/s.<br/>total: 186.15 ± 1.10ms. best: 185.11ms.<br/>avg: 3.27 ± 0.02ms<br/>min: 0.13 ± 0.01ms<br/>max: 18.91 ± 0.03ms<br/>p99: 18.91 ± 0.03ms<br/><br/>500 matches, radius=10<br/>ops: 108.66 ± 0.38 ops/s. best: 109.13ops/s.<br/>total: 589.03 ± 2.07ms. best: 586.43ms.<br/>avg: 9.20 ± 0.03ms<br/>min: 0.14 ± 0.00ms<br/>max: 109.95 ± 0.61ms<br/>p99: 109.95 ± 0.61ms<br/><br/>500 matches, radius=20<br/>ops: 22.13 ± 0.03 ops/s. best: 22.18ops/s.<br/>total: 2937.06 ± 4.02ms. best: 2930.66ms.<br/>avg: 45.19 ± 0.06ms<br/>min: 0.20 ± 0.00ms<br/>max: 574.14 ± 1.42ms<br/>p99: 574.14 ± 1.42ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>ops: 24297.05 ± 71.26 ops/s. best: 24396.80ops/s.<br/>total: 411.58 ± 1.27ms. best: 409.89ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.18 ± 0.05ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18610.96 ± 28.59 ops/s. best: 18665.96ops/s.<br/>total: 537.32 ± 0.82ms. best: 535.73ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.16 ± 0.01ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14770.66 ± 31.32 ops/s. best: 14819.14ops/s.<br/>total: 677.02 ± 1.43ms. best: 674.80ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.14 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 24515.17 ± 74.68 ops/s. best: 24597.12ops/s.<br/>total: 407.92 ± 1.25ms. best: 406.55ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.20 ± 0.06ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18903.61 ± 32.13 ops/s. best: 18975.22ops/s.<br/>total: 529.00 ± 0.91ms. best: 527.00ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14966.77 ± 13.95 ops/s. best: 14985.57ops/s.<br/>total: 668.15 ± 0.62ms. best: 667.31ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.13 ± 0.00ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>ops: 24157.43 ± 366.67 ops/s. best: 24491.45ops/s.<br/>total: 414.11 ± 6.60ms. best: 408.31ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.21 ± 0.06ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18680.17 ± 32.77 ops/s. best: 18736.28ops/s.<br/>total: 535.33 ± 0.85ms. best: 533.72ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.23 ± 0.07ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14747.39 ± 28.97 ops/s. best: 14793.76ops/s.<br/>total: 678.09 ± 1.33ms. best: 675.96ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.19 ± 0.02ms<br/>p99: 0.14 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 24744.34 ± 137.43 ops/s. best: 24954.07ops/s.<br/>total: 404.15 ± 2.20ms. best: 400.74ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.05ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18995.84 ± 49.81 ops/s. best: 19103.32ops/s.<br/>total: 526.44 ± 1.38ms. best: 523.47ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14952.20 ± 38.61 ops/s. best: 15015.90ops/s.<br/>total: 668.80 ± 1.77ms. best: 665.96ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.19 ± 0.01ms<br/>p99: 0.13 ± 0.00ms |
| random_route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 460.77 ± 0.72 ops/s. best: 461.90ops/s.<br/>total: 2135.56 ± 3.34ms. best: 2130.33ms.<br/>avg: 2.17 ± 0.00ms<br/>min: 0.32 ± 0.01ms<br/>max: 3.90 ± 0.17ms<br/>p99: 3.18 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 531.68 ± 3.31 ops/s. best: 535.17ops/s.<br/>total: 1880.92 ± 11.78ms. best: 1868.57ms.<br/>avg: 1.88 ± 0.01ms<br/>min: 0.06 ± 0.00ms<br/>max: 4.66 ± 0.12ms<br/>p99: 3.85 ± 0.05ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 876.59 ± 9.28 ops/s. best: 896.09ops/s.<br/>total: 1122.69 ± 11.76ms. best: 1098.10ms.<br/>avg: 1.14 ± 0.01ms<br/>min: 0.27 ± 0.00ms<br/>max: 1.84 ± 0.03ms<br/>p99: 1.64 ± 0.03ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 968.24 ± 7.84 ops/s. best: 978.69ops/s.<br/>total: 1032.89 ± 8.40ms. best: 1021.78ms.<br/>avg: 1.03 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 2.72 ± 0.01ms<br/>p99: 2.24 ± 0.03ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 477.38 ± 6.33 ops/s. best: 486.34ops/s.<br/>total: 2061.75 ± 27.43ms. best: 2023.28ms.<br/>avg: 2.10 ± 0.03ms<br/>min: 0.31 ± 0.01ms<br/>max: 3.79 ± 0.24ms<br/>p99: 3.04 ± 0.05ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 545.00 ± 5.18 ops/s. best: 553.22ops/s.<br/>total: 1835.08 ± 17.09ms. best: 1807.60ms.<br/>avg: 1.84 ± 0.02ms<br/>min: 0.06 ± 0.01ms<br/>max: 4.66 ± 0.19ms<br/>p99: 3.90 ± 0.06ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 912.08 ± 1.62 ops/s. best: 914.60ops/s.<br/>total: 1078.85 ± 1.89ms. best: 1075.88ms.<br/>avg: 1.10 ± 0.00ms<br/>min: 0.27 ± 0.00ms<br/>max: 1.78 ± 0.02ms<br/>p99: 1.55 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 989.55 ± 12.25 ops/s. best: 1002.87ops/s.<br/>total: 1010.79 ± 12.72ms. best: 997.14ms.<br/>avg: 1.01 ± 0.01ms<br/>min: 0.05 ± 0.01ms<br/>max: 2.87 ± 0.04ms<br/>p99: 2.19 ± 0.04ms |
| random_route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 215.00 ± 2.10 ops/s. best: 219.42ops/s.<br/>total: 4577.21 ± 44.30ms. best: 4484.48ms.<br/>avg: 4.65 ± 0.05ms<br/>min: 0.34 ± 0.01ms<br/>max: 10.13 ± 0.15ms<br/>p99: 7.72 ± 0.03ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 209.28 ± 1.43 ops/s. best: 211.16ops/s.<br/>total: 4778.66 ± 32.69ms. best: 4735.82ms.<br/>avg: 4.78 ± 0.03ms<br/>min: 0.05 ± 0.00ms<br/>max: 11.37 ± 0.66ms<br/>p99: 10.02 ± 0.11ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 292.27 ± 2.40 ops/s. best: 295.75ops/s.<br/>total: 3367.02 ± 27.61ms. best: 3327.16ms.<br/>avg: 3.42 ± 0.03ms<br/>min: 0.29 ± 0.01ms<br/>max: 7.95 ± 0.03ms<br/>p99: 5.81 ± 0.09ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 263.21 ± 1.08 ops/s. best: 264.97ops/s.<br/>total: 3799.33 ± 15.59ms. best: 3774.06ms.<br/>avg: 3.80 ± 0.02ms<br/>min: 0.04 ± 0.00ms<br/>max: 9.02 ± 1.20ms<br/>p99: 7.55 ± 0.04ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 226.58 ± 0.88 ops/s. best: 227.99ops/s.<br/>total: 4342.93 ± 16.84ms. best: 4315.95ms.<br/>avg: 4.41 ± 0.02ms<br/>min: 0.32 ± 0.01ms<br/>max: 9.51 ± 0.15ms<br/>p99: 7.53 ± 0.16ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 208.77 ± 2.41 ops/s. best: 212.10ops/s.<br/>total: 4790.81 ± 55.56ms. best: 4714.78ms.<br/>avg: 4.79 ± 0.06ms<br/>min: 0.05 ± 0.00ms<br/>max: 11.49 ± 0.36ms<br/>p99: 10.19 ± 0.38ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 287.39 ± 1.87 ops/s. best: 290.37ops/s.<br/>total: 3424.14 ± 22.32ms. best: 3388.74ms.<br/>avg: 3.48 ± 0.02ms<br/>min: 0.29 ± 0.00ms<br/>max: 8.24 ± 0.15ms<br/>p99: 5.98 ± 0.09ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 260.42 ± 0.97 ops/s. best: 262.00ops/s.<br/>total: 3839.99 ± 14.24ms. best: 3816.83ms.<br/>avg: 3.84 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 8.76 ± 0.69ms<br/>p99: 7.66 ± 0.09ms |
| random_table_ch | 250 tables, 3 coordinates<br/>ops: 1430.32 ± 15.83 ops/s. best: 1446.93ops/s.<br/>total: 174.82 ± 2.03ms. best: 172.78ms.<br/>avg: 0.70 ± 0.01ms<br/>min: 0.47 ± 0.01ms<br/>max: 1.17 ± 0.32ms<br/>p99: 0.93 ± 0.04ms<br/><br/>250 tables, 25 coordinates<br/>ops: 165.85 ± 0.38 ops/s. best: 166.67ops/s.<br/>total: 1507.40 ± 3.41ms. best: 1500.01ms.<br/>avg: 6.03 ± 0.01ms<br/>min: 5.26 ± 0.02ms<br/>max: 6.95 ± 0.22ms<br/>p99: 6.65 ± 0.03ms<br/><br/>250 tables, 50 coordinates<br/>ops: 81.51 ± 0.36 ops/s. best: 82.12ops/s.<br/>total: 3067.01 ± 13.32ms. best: 3044.28ms.<br/>avg: 12.27 ± 0.05ms<br/>min: 11.26 ± 0.08ms<br/>max: 13.28 ± 0.11ms<br/>p99: 13.16 ± 0.08ms | 250 tables, 3 coordinates<br/>ops: 1425.49 ± 8.07 ops/s. best: 1432.00ops/s.<br/>total: 175.39 ± 1.01ms. best: 174.58ms.<br/>avg: 0.70 ± 0.00ms<br/>min: 0.51 ± 0.00ms<br/>max: 1.11 ± 0.33ms<br/>p99: 0.89 ± 0.02ms<br/><br/>250 tables, 25 coordinates<br/>ops: 166.63 ± 0.46 ops/s. best: 167.25ops/s.<br/>total: 1500.35 ± 4.16ms. best: 1494.78ms.<br/>avg: 6.00 ± 0.02ms<br/>min: 5.30 ± 0.01ms<br/>max: 7.16 ± 0.61ms<br/>p99: 6.70 ± 0.28ms<br/><br/>250 tables, 50 coordinates<br/>ops: 82.54 ± 0.31 ops/s. best: 83.19ops/s.<br/>total: 3028.96 ± 11.51ms. best: 3005.24ms.<br/>avg: 12.12 ± 0.05ms<br/>min: 11.31 ± 0.05ms<br/>max: 13.17 ± 0.05ms<br/>p99: 12.88 ± 0.05ms |
…